experimenting with single type parser, finished systable.RowToData

+++ src/dbsystem/hgtype/rawtype/args.go
@@ -3,14 +3,14 @@ package rawtype
 import "HomegrownDB/common/bparse"
 
 type Args struct {
-	Length   uint32
+	Length   int
 	Nullable bool
 	VarLen   bool
 	UTF8     bool
 }
 
 func SerializeArgs(args Args, s *bparse.Serializer) {
-	s.Uint32(args.Length)
+	s.Uint64(uint64(args.Length))
 	s.Bool(args.Nullable)
 	s.Bool(args.VarLen)
 	s.Bool(args.UTF8)
@@ -18,7 +18,7 @@ func SerializeArgs(args Args, s *bparse.Serializer) {
 
 func DeserializeArgs(d *bparse.Deserializer) Args {
 	return Args{
-		Length:   d.Uint32(),
+		Length:   int(d.Uint64()),
 		Nullable: d.Bool(),
 		VarLen:   d.Bool(),
 		UTF8:     d.Bool(),
